In progress

+++ UCT.cpp
@@ -24,7 +24,7 @@ std::mutex treePolicyMutex;
 std::mutex backUpMutex;
 
 double diffclock(clock_t clock1, clock_t clock2) {
-  double diffticks = clock1 - clock2;
+  double diffticks = double(clock1 - clock2);
   double diffms = diffticks / (CLOCKS_PER_SEC / 1000);
   return diffms;
 }
